refactor: separated UserService into Create/List UserService
